+++ StorybookInteractive/Assets.xcassets/BackgroundBook1Page7.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "BackgroundBook1Page7.png",
+      "filename" : "page71.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
