+++ StorybookInteractive/Assets.xcassets/BackgroundBook1Page7WithCharacter.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "BackgroundBook1Page7WithCharacter.png",
+      "filename" : "Page7.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
