--- StorybookInteractive/PhysicalCardClassifier.mlproj/Checkpoints/09DAC466-1824-421B-AC9A-F7DD3CA14204.AllCardClassifier.checkpoint/training-000000075.checkpoint/model.pkg/optimizer.json
@@ -1,1 +1,0 @@
-[{"stepSize":0.77862262725830078,"point":{"strides":[1],"scalar":"double","shape":[14343],"index":0},"stopped":false,"hessian":{"shape":[0,0],"scalars":[]},"gradient":{"strides":[1],"scalar":"double","shape":[14343],"index":1},"value":0.002075283443211878,"residual":1.565988242578392}]
