--- StorybookInteractive/PhysicalCardClassifier.mlproj/Checkpoints/09DAC466-1824-421B-AC9A-F7DD3CA14204.AllCardClassifier.checkpoint/training-000000075.checkpoint/model.pkg/pipeline.json
@@ -1,1 +1,0 @@
-[{"model":"4B5BC945-4A0E-40A7-92B4-56EA855DA352","weights":{"strides":[1],"scalar":"double","shape":[14343],"index":0}},["Bantal","Bebek","Bulu","Bunga","Burung","Matahari","NonClassified","Roti"]]
