--- StorybookInteractive/PhysicalCardClassifier.mlproj/Checkpoints/D4FF5FFE-D1D3-47AC-BD86-0A3C6DAD74C5.PhysicalCardClassifier1.checkpoint/training-000000025.checkpoint/model.pkg/optimizer.json
@@ -1,1 +1,0 @@
-[{"stepSize":0.83709537982940674,"point":{"strides":[1],"scalar":"double","shape":[6147],"index":0},"stopped":false,"hessian":{"shape":[0,0],"scalars":[]},"gradient":{"strides":[1],"scalar":"double","shape":[6147],"index":1},"value":0.00055241150369835521,"residual":0.11496545408525458}]
