--- StorybookInteractive/PhysicalCardClassifier.mlproj/Checkpoints/D4FF5FFE-D1D3-47AC-BD86-0A3C6DAD74C5.PhysicalCardClassifier1.checkpoint/training-000000025.checkpoint/model.pkg/pipeline.json
@@ -1,1 +1,0 @@
-[{"model":"CA9E4F55-2FAA-4F26-9C9C-47A29CBA1C4A","weights":{"strides":[1],"scalar":"double","shape":[6147],"index":0}},["Bebek","Burung","NonClassified","Roti"]]
