--- StorybookInteractive/PhysicalCardClassifier.mlproj/Data Sources/TrainingData.json
@@ -1,1 +1,0 @@
-{"experimentIDs":["D4FF5FFE-D1D3-47AC-BD86-0A3C6DAD74C5","09DAC466-1824-421B-AC9A-F7DD3CA14204"],"title":"TrainingData","dataResult":{"itemsCount":856,"dataURL":"file:\/\/\/Users\/juan\/Documents\/TrainingData\/","classesCount":8,"dataAnalysis":{"summary":{"canUseValidation":1,"classCount":8,"sampleCount":856},"classifierAnalysis":{"labelStats":{"Burung":87,"NonClassified":219,"Bunga":82,"Bulu":74,"Matahari":107,"Roti":96,"Bebek":115,"Bantal":76}},"issues":[{"description":"8 classes found, including NonClassified, Bebek, Matahari","severity":"debug"},{"description":"856 images found, automatically creating validation set","severity":"good"},{"description":"Smallest class is \"Bulu\" with 74 images","severity":"debug"},{"severity":"debug","description":"Largest class is \"NonClassified\" with 219 images"}]},"dataSourceID":"70A5DBB9-C351-495A-BA38-63378C821784","createdDate":"2024-08-19T03:41:30Z"},"createdDate":"2024-08-19T03:41:30Z","id":"70A5DBB9-C351-495A-BA38-63378C821784","dataURL":"file:\/\/\/Users\/juan\/Documents\/TrainingData\/","preferredFilename":"TrainingData","dataPathRelativeToProjectURL":"..\/..\/..\/TrainingData\/","activityEntities":["dataSourceCreated",[{"event":"dataSourceCreated","title":"Data Source Created","subtitle":"TrainingData","createdDate":"2024-08-19T03:41:30Z","id":"F2D2BFE2-67A4-4841-9BA5-6118AB1F847D"}],"dataSourceTrainingStarted",[{"id":"A2565245-9849-472E-BEC0-589EA011F2E9","title":"Used By","createdDate":"2024-08-19T03:41:35Z","event":"dataSourceTrainingStarted","subtitle":"PhysicalCardClassifier 1"},{"id":"0B951E5B-9F52-4B4D-A7E4-FC0B8EBC957D","createdDate":"2024-08-20T17:19:52Z","event":"dataSourceTrainingStarted","title":"Used By","subtitle":"AllCardClassifier"},{"event":"dataSourceTrainingStarted","createdDate":"2024-08-20T17:20:13Z","subtitle":"AllCardClassifier","id":"A7E1F8B7-38FF-46E9-B05C-434FFBA8D2D3","title":"Used By"},{"event":"dataSourceTrainingStarted","subtitle":"AllCardClassifier","id":"00550CEC-2306-4950-A30F-F2B986F2AFBB","title":"Used By","createdDate":"2024-08-20T17:20:18Z"}],"dataSourceMetrics",[{"id":"333C10D6-7166-4CEC-B24B-D617C8C8D0D0","event":"dataSourceMetrics","subtitle":"4 Classes, 353 Items","title":"Data Summary","createdDate":"2024-08-19T03:41:35Z"},{"event":"dataSourceMetrics","title":"Data Summary","createdDate":"2024-08-20T17:19:52Z","subtitle":"8 Classes, 856 Items","id":"C06E6C9E-6947-4F3A-B2AF-70AB320D9DB1"},{"createdDate":"2024-08-20T17:20:13Z","title":"Data Summary","event":"dataSourceMetrics","id":"B531699F-CAA8-46BA-BFFC-4B9F9C3D2264","subtitle":"8 Classes, 856 Items"},{"event":"dataSourceMetrics","subtitle":"8 Classes, 856 Items","id":"EC51C1A6-CADB-4D85-9EB2-993A5A285FF1","title":"Data Summary","createdDate":"2024-08-20T17:20:18Z"}]],"version":"1.0","pathType":0,"modifiedDate":"2024-08-20T17:20:18Z"}
