--- StorybookInteractive/PhysicalCardClassifier.mlproj/Project.json
@@ -1,1 +1,0 @@
-{"authorName":"Nathanael Juan Gauthama","modifiedDate":"2024-08-19T03:41:21Z","version":"1.1","createdDate":"2024-08-19T03:41:21Z","location":"","settingsEntity":{"snapshotOnResumeTraining":false,"snapshotOnExtendTraining":true},"license":"","id":"B20D5723-245F-496D-A060-3034C274096B","projectName":"PhysicalCardClassifier","title":"Image Classification","taskType":"imageClassifier","versionErrorTitle":"Project file cannot be opened.","description":"","activityEntities":["projectCreated",[{"id":"2D757778-4A8D-4A24-B503-5666CC277A9B","title":"Project Created","createdDate":"2024-08-19T03:41:21Z","event":"projectCreated","subtitle":"PhysicalCardClassifier"}]],"versionErrorSubtitle":"You need a newer version of Create ML to open this file. Download the latest version."}
