--- StorybookInteractive/Presentation/Views/ButtonComponent.swift
@@ -1,35 +1,0 @@
-//
-//  ButtonComponent.swift
-//  StorybookInteractive
-//
-//  Created by Nathanael Juan Gauthama on 15/08/24.
-//
-
-import UIKit
-
-class ButtonComponent: UIButton {
-    
-
-}
-
-extension UIImage {
-  public static func pixel(ofColor color: UIColor) -> UIImage {
-    let pixel = CGRect(x: 0.0, y: 0.0, width: 1.0, height: 1.0)
-
-    UIGraphicsBeginImageContext(pixel.size)
-    defer { UIGraphicsEndImageContext() }
-
-    guard let context = UIGraphicsGetCurrentContext() else { return UIImage() }
-
-    context.setFillColor(color.cgColor)
-    context.fill(pixel)
-
-    return UIGraphicsGetImageFromCurrentImageContext() ?? UIImage()
-  }
-}
-
-extension UIButton {
-    func setBackgroundColor(_ backgroundColor: UIColor, for state: UIControl.State) {
-        self.setBackgroundImage(.pixel(ofColor: backgroundColor), for: state)
-    }
-}
